Consolidating ConsentServiceTest and ConsentServiceMockTest to ConsentServiceTest.
